requirements.txt
airbnb==2.3.2
argcomplete==1.12.0
asgiref==3.2.10
attrs==20.3.0
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.3.3
blivet-gui==2.3.0
Brlapi==0.8.2
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
cryptography==3.4.6
cupshelpers==1.0
dasbus==1.4
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.2
distro==1.5.0
Django==3.1.13
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.28.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
fros==1.1
gpg==1.15.1
humanize==0.5.1
idna==2.10
jsonschema==3.2.0
langtable==0.0.56
libcomps==0.1.17
lxml==4.6.3
Mako==1.1.4.dev0
MarkupSafe==1.1.1
mod-wsgi==4.7.1
mysqlclient==2.0.3
nftables==0.1
olefile==0.46
packaging==20.9
paramiko==2.7.2
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.1.2
pipenv==2021.5.29
platformdirs==2.3.0
ply==3.11
productmd==1.33
ptyprocess==0.6.0
pwquality==1.4.4
pyasn1==0.4.8
pycairo==1.20.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.1
PyGObject==3.40.1
pykickstart==3.32
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyparted==3.11.7
pyrsistent==0.17.3
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-meh==0.50
pytz==2021.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
requests==2.25.1
requests-file==1.5.1
requests-ftp==0.3.1
rpm==4.16.1.3
selinux==3.2
sepolicy==3.2
setools==4.4.0
simpleaudio==1.0.4
simpleline==1.8
six==1.15.0
slip==0.6.4
slip.dbus==0.6.4
sos==4.1
soupsieve==2.2
sqlparse==0.4.1
systemd-python==234
Tempita==0.5.1
texttable==1.6.2
urllib3==1.25.10
virtualenv==20.7.2
virtualenv-clone==0.5.6
websocket-client==0.57.0
